moved output serialization structs to api
