+++ core/cli/serialization.go
@@ -3,7 +3,6 @@ package cli
 import (
 	"encoding/json"
 	"io"
-	"time"
 
 	"go.polydawn.net/repeatr/api/def"
 
@@ -13,31 +12,13 @@ import (
 
 var _ io.Writer = &journalSerializer{}
 
-type ctxPairs []interface{}
-
-func (ctxPairs) MapBySlice() {}
-
-type logItem struct {
-	Level int       `json:"lvl"`
-	Msg   string    `json:"msg"`
-	Ctx   ctxPairs  `json:"ctx"`
-	Time  time.Time `json:"t"`
-}
-
-type serializedOutput struct {
-	RunID     def.RunID      `json:"runID,omitempty"`
-	RunRecord *def.RunRecord `json:"runRecord,omitempty"`
-	Journal   string         `json:"journal,omitempty"`
-	Log       *logItem       `json:"log,omitempty"`
-}
-
 type journalSerializer struct {
 	Delegate io.Writer
 	RunID    def.RunID
 }
 
 func (a *journalSerializer) Write(b []byte) (int, error) {
-	jo := serializedOutput{Journal: string(b), RunID: a.RunID}
+	jo := def.SerializedOutput{Journal: string(b), RunID: a.RunID}
 	out, err := json.Marshal(jo)
 	if err != nil {
 		panic(err)
@@ -47,7 +28,7 @@ func (a *journalSerializer) Write(b []byte) (int, error) {
 }
 
 func serializeRunRecord(wr io.Writer, runID def.RunID, rr *def.RunRecord) error {
-	err := codec.NewEncoder(wr, &codec.JsonHandle{}).Encode(serializedOutput{
+	err := codec.NewEncoder(wr, &codec.JsonHandle{}).Encode(def.SerializedOutput{
 		RunRecord: rr,
 		RunID:     runID,
 	})
@@ -57,13 +38,13 @@ func serializeRunRecord(wr io.Writer, runID def.RunID, rr *def.RunRecord) error 
 
 func logHandler(wr io.Writer) log15.Handler {
 	h := log15.FuncHandler(func(r *log15.Record) error {
-		li := &logItem{
+		li := &def.LogItem{
 			Level: int(r.Lvl),
 			Msg:   r.Msg,
 			Ctx:   r.Ctx,
 			Time:  r.Time,
 		}
-		err := codec.NewEncoder(wr, &codec.JsonHandle{}).Encode(serializedOutput{Log: li})
+		err := codec.NewEncoder(wr, &codec.JsonHandle{}).Encode(def.SerializedOutput{Log: li})
 		wr.Write([]byte{'\n'})
 		return err
 	})
